src/backend/requirements.txt
aioredis==1.3.1
appnope==0.1.2
argcomplete==1.12.3
argon2-cffi==21.1.0
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
autopep8==1.5.7
awsebcli==3.20.2
backcall==0.2.0
bcrypt==3.2.0
bleach==4.1.0
blessed==1.19.0
boto3==1.18.0
botocore==1.21.65
CacheControl==0.12.6
cached-property==1.5.2
cachetools==4.2.2
cement==2.8.2
certifi==2021.5.30
cffi==1.14.6
channels==3.0.4
channels-redis==3.3.0
charset-normalizer==2.0.6
colorama==0.4.3
constantly==15.1.0
cryptography==3.4.8
daphne==3.0.2
debugpy==1.4.3
decorator==5.1.0
defusedxml==0.7.1
Django==3.2.7
django-cors-headers==3.8.0
django-environ==0.8.1
django-extensions==3.1.3
django-storages==1.12.3
djangorestframework==3.12.4
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
entrypoints==0.3
et-xmlfile==1.1.0
firebase-admin==5.0.3
future==0.16.0
google-api-core==2.0.1
google-api-python-client==2.22.0
google-auth==2.1.0
google-auth-httplib2==0.1.0
google-cloud-core==2.0.0
google-cloud-firestore==2.3.3
google-cloud-secret-manager==2.7.2
google-cloud-storage==1.42.2
google-crc32c==1.2.0
google-resumable-media==2.0.3
googleapis-common-protos==1.53.0
grpc-google-iam-v1==0.12.3
grpcio==1.40.0
gunicorn==20.1.0
hiredis==2.0.0
httplib2==0.19.1
hyperlink==21.0.0
idna==3.2
importlib-metadata==4.8.1
incremental==21.3.0
install==1.3.4
ipykernel==6.4.1
ipython==7.27.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.1
jmespath==0.10.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.0.3
jupyter-console==6.4.0
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
libcst==0.3.21
lorem==0.1.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
msgpack==1.0.2
mypy-extensions==0.4.3
names==0.3.0
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.4
numpy==1.21.2
openpyxl==3.0.9
packaging==21.0
pandas==1.3.3
pandocfilters==1.5.0
paramiko==2.8.1
parso==0.8.2
pathspec==0.5.9
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
prometheus-client==0.11.0
prompt-toolkit==3.0.20
proto-plus==1.19.0
protobuf==3.18.0
psycopg2-binary==2.9.2
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
Pygments==2.10.0
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python-http-client==3.3.2
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.11.2
requests==2.26.0
rsa==4.7.2
s3transfer==0.5.0
semantic-version==2.8.5
Send2Trash==1.8.0
sendgrid==6.8.1
service-identity==21.1.0
six==1.14.0
sqlparse==0.4.2
starkbank-ecdsa==1.1.1
stripe==2.60.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
texttable==1.6.4
toml==0.10.2
tornado==6.1
traitlets==5.1.0
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.2
typing-inspect==0.7.1
uritemplate==3.0.1
urllib3==1.26.5
watchtower==3.0.0
wcwidth==0.1.9
webencodings==0.5.1
websocket-client==0.59.0
widgetsnbextension==3.5.1
zipp==3.5.0
zope.interface==5.4.0
